+++ packages/adguard-api-mv3/src/background/request-blocking-logger.ts
@@ -40,6 +40,8 @@ export type RequestBlockingEvent = {
     assumedFilterId: number;
     /** Request mime type. */
     requestType: ContentType;
+    /** Tracker category ID. */
+    trackerCategoryId: number;
 };
 
 export interface RequestBlockingLoggerInterface {
@@ -97,6 +99,7 @@ export class RequestBlockingLogger implements RequestBlockingLoggerInterface {
             filterId,
             ruleIndex,
             isAllowlist,
+            companyCategory,
         } = event.data;
 
         // exclude allowlist rules
@@ -111,6 +114,8 @@ export class RequestBlockingLogger implements RequestBlockingLoggerInterface {
             requestUrl,
             referrerUrl: frameUrl,
             requestType,
+            // always provided for MV3
+            trackerCategoryId: companyCategory!,
         });
     }
 }