+++ packages/tswebextension/CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version changes -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## [Unreleased]
+
+### Added
+
+- New `companies` CLI command for downloading companies database from `AdguardTeam/companiesdb` repository.
+
+### Changed
+
+- Add optional `companiesDbPath` argument to `TsWebExtension` constructor for MV3 to specify the path to the companies database file.
+- filtering log `ApplyBasicRuleEvent` now contains `companyCategory` id from `AdguardTeam/companiesdb` database in MV3 build. This update allows determinate the tracker category for the blocked request.
+
+
 ## [2.0.0-alpha.2] - 2024-07-17
 
 ### Changed